refactor: create hooks.ts file
