+++ src/lib/substrate/index.ts
@@ -1,6 +1,9 @@
 import utils from './utils';
 
+export * from './context/hooks';
 export * from './context/provider';
 export * from './context/types';
 
+// ray test touch <
 export { utils as substrateUtils };
+// ray test touch >
